Move Arc Types into separate file and fix imports
