+++ arc-mlir/src/CMakeLists.txt
@@ -14,8 +14,7 @@ set(TBLGEN_INCDIRS "-I${CMAKE_CURRENT_SOURCE_DIR}/../llvm-project/mlir/include/"
 mlir_tablegen(Opts.inc -gen-rewriters ${TBLGEN_INCDIRS} "-I${CMAKE_CURRENT_SOURCE_DIR}/include")
 add_public_tablegen_target(ArcDialectOptsIncGen)
 
-
-set(ARCMLIR_MODULES Main.cpp lib/Dialect.cpp lib/Opts.cpp)
+set(ARCMLIR_MODULES Main.cpp lib/Dialect.cpp lib/Opts.cpp lib/Types.cpp)
 
 set(LIBS
   MLIRAnalysis
