+++ arc-mlir/src/include/arc/Dialect.h
@@ -23,19 +23,15 @@
 #ifndef ARC_DIALECT_H_
 #define ARC_DIALECT_H_
 
-#include "mlir/IR/StandardTypes.h"
+#include "arc/Types.h"
+#include <mlir/IR/Builders.h>
 #include <mlir/IR/Dialect.h>
 #include <mlir/IR/Function.h>
-#include <mlir/IR/Builders.h>
 #include <mlir/IR/Operation.h>
 
 using namespace mlir;
 
 namespace arc {
-
-namespace detail {
-struct AppenderTypeStorage;
-};
 
 /// This is the definition of the Arc dialect.
 class ArcDialect : public mlir::Dialect {
@@ -55,34 +51,6 @@ public:
 /// arc operations.
 #define GET_OP_CLASSES
 #include "arc/Ops.h.inc"
-
-//===----------------------------------------------------------------------===//
-// Toy Types
-//===----------------------------------------------------------------------===//
-
-namespace types {
-
-enum Kind {
-  // These kinds will be used by Arc.
-  Appender = Type::Kind::FIRST_PRIVATE_EXPERIMENTAL_0_TYPE,
-};
-
-bool isValueType(Type type);
-
-bool isBuilderType(Type type);
-
-class AppenderType
-    : public Type::TypeBase<AppenderType, Type, detail::AppenderTypeStorage> {
-public:
-  using Base::Base;
-
-  static bool kindof(unsigned kind) { return kind == arc::types::Appender; }
-
-  static AppenderType get(Type mergeType);
-
-  Type getMergeType();
-};
-} // namespace types
 
 } // end namespace arc
